+++ pkg/jwt/util_test.go
@@ -1,14 +1,11 @@
-package jwt
+package jwt_test
 
 import (
 	"context"
-	"crypto/rsa"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"reflect"
-	"strings"
-	"time"
 )
 
 type testData struct {
@@ -67,89 +64,18 @@ func hasInContext(wantKey interface{}, wantValue interface{}) checkReqFunc {
 	}
 }
 
-// hasKeyInContext returns a function that checks that the context of the request contains
-// the expected key with a non-nil value.
-func hasKeyInContext(wantKey interface{}) checkReqFunc {
+// hasInContextExtract returns a function that allows to extract some value from the context of the request
+// and compare it with the expected value.
+func hasInContextExtract(extract func(context.Context) (interface{}, error), wantValue interface{}) checkReqFunc {
 	return func(r *http.Request) error {
-		haveValue := r.Context().Value(wantKey)
-		if haveValue == nil {
-			return fmt.Errorf("expected to find key %v; have nil", wantKey)
-		}
-
-		return nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// HTTP Request Builder and OKHandler
-////////////////////////////////////////////////////////////////////////////////
-
-type requestBuilder func(*http.Request) error
-
-func buildRequest(url string, fns ...requestBuilder) *http.Request {
-	r := httptest.NewRequest("", url, nil)
-
-	for _, fn := range fns {
-		_ = fn(r)
-	}
-
-	return r
-}
-
-func withAuthHeader(key *rsa.PrivateKey, options ...TokenOption) requestBuilder {
-	return func(r *http.Request) error {
-		options = append(
-			// add a default expiration time as the token is not valid without one
-			[]TokenOption{WithExpirationTime(time.Now().Add(1 * time.Minute))},
-			options...,
-		)
-		t, err := CreateAccessToken(key, options...)
+		haveValue, err := extract(r.Context())
 		if err != nil {
-			return err
+			return fmt.Errorf("can't extract from context: %w", err)
 		}
-
-		r.Header.Add("Authorization", "Bearer "+t.AccessToken)
+		if !reflect.DeepEqual(wantValue, haveValue) {
+			return fmt.Errorf("unexpected context value found in context; want: %v, have: %v", wantValue, haveValue)
+		}
 
 		return nil
 	}
-}
-
-func withOwnerPath(owner string) string {
-	var builder strings.Builder
-
-	builder.WriteString("/users/")
-	builder.WriteString(owner)
-	builder.WriteString("/records/456")
-
-	return builder.String()
-}
-
-func withOwnerURL(owner string) string {
-	var builder strings.Builder
-
-	builder.WriteString("http://test.data4life.care")
-	builder.WriteString(withOwnerPath(owner))
-
-	return builder.String()
-}
-
-func okHandler(w http.ResponseWriter, r *http.Request) {}
-
-////////////////////////////////////////////////////////////////////////////////
-// Test Logger
-////////////////////////////////////////////////////////////////////////////////
-
-type testLogger struct{}
-
-func (testLogger) ErrUserAuth(ctx context.Context, err error) error {
-	fmt.Println(err)
-	return nil
-}
-func (testLogger) InfoGeneric(ctx context.Context, msg string) error {
-	fmt.Println(msg)
-	return nil
-}
-func (testLogger) ErrGeneric(ctx context.Context, err error) error {
-	fmt.Println(err)
-	return nil
 }